+++ lib/mutation_tester/fork_runner.rb
@@ -153,7 +153,8 @@ module MutationTester
       @ready
     end
 
-    def execute(spec_file, timeout: nil, chdir: nil, capture: false, args: [], stop_on_first_failure: false)
+    def execute(spec_file, timeout: nil, chdir: nil, capture: false, args: [], stop_on_first_failure: false,
+                mirror_of: nil)
       log = capture ? Tempfile.new(['mutation_tester_fork', '.log']) : nil
       job = {
         spec: spec_file,
@@ -161,7 +162,8 @@ module MutationTester
         chdir: chdir,
         log: log&.path,
         args: args,
-        stop_on_first_failure: stop_on_first_failure
+        stop_on_first_failure: stop_on_first_failure,
+        mirror_of: mirror_of
       }
       @job_writer.puts(JSON.generate(job))
       status = await_result(timeout)['status']